requirements.txt
torch==1.13.1
torchvision==0.14.1
transformers==4.26.0
opencv-python==4.6.0.66
scikit-learn==1.1.3
Pillow==9.3.0
numpy==1.23.5
